Use size_t instead of int for remote file size.
git-svn-id: svn+ssh://cvs.lpds.sztaki.hu/var/lib/svn/szdg/dcapi/trunk@2407 a7169a2c-3604-0410-bc95-c702d8d87f7a

+++ dcapi/include/dc_common.h
@@ -21,6 +21,7 @@ extern "C" {
 #include <sys/syslog.h>
 #endif
 
+#include <string.h>
 #include <stdarg.h>
 
 /********************************************************************
@@ -98,7 +99,7 @@ struct _DC_RemoteFile
 	char			*label;
 	char			*url;
 	char			*remotefilehash;
-	int			remotefilesize;
+	size_t			remotefilesize;
 };
